修改信息

+++ src/main/java/com/acc/controller/BxProductController.java
@@ -302,11 +302,14 @@ public class BxProductController {
             if(bxProduct!=null){
                 if (bxProductVideo != null) {
                     if(file != null){
-                        BxProductVideo oldbxProductVideo = bxProductService.getProductDetailVideoById(String.valueOf(bxProductVideo.getId()));
                         String path = (String)request.getSession().getServletContext().getAttribute("proRoot");
-                        String fileSavePath=path + Constants.proVideoPath + oldbxProductVideo.getProductId() + "/";
-                        //删除old
-                        new File(fileSavePath+oldbxProductVideo.getVideoUrl()).delete();
+                        String fileSavePath=path + Constants.proVideoPath + bxProductVideo.getProductId() + "/";
+                        BxProductVideo oldbxProductVideo = bxProductService.getProductDetailVideoById(String.valueOf(bxProductVideo.getId()));
+                        if(oldbxProductVideo!=null){
+                            //删除old
+                            new File(fileSavePath+oldbxProductVideo.getVideoUrl()).delete();
+                            bxProductService.deleteProductDetailVideoById(String.valueOf(bxProductVideo.getId()));
+                        }
                         //添加new
                         Map<String,Object> mapImg = PictureChange.imageUpload(file,fileSavePath,true,false);
                         int re = Integer.valueOf((String)mapImg.get("code")).intValue();
@@ -314,7 +317,6 @@ public class BxProductController {
                             List<String> videoNameList = (List<String>)mapImg.get("list");
                             if(videoNameList!=null && videoNameList.size()>0){
                                 bxProductVideo.setVideoUrl(videoNameList.get(0));
-                                bxProductService.deleteProductDetailVideoById(String.valueOf(bxProductVideo.getId()));
                                 bxProductService.insertProductVideo(bxProductVideo);
                             }
                             result = "添加成功";
